#810 - mu: break out core async into separate directory
