+++ src/mu/async_/mod.rs
@@ -1,0 +1,5 @@
+//  SPDX-FileCopyrightText: Copyright 2022 James M. Putnam (putnamjm.design@gmail.com)
+//  SPDX-License-Identifier: MIT
+
+//! async module
+pub mod context;
