+++ src/mu/core/reader.rs
@@ -15,9 +15,9 @@ use crate::{
         char::Char,
         fixnum::Fixnum,
         float::Float,
-        r#struct::{Core as _, Struct},
         stream::{Core as _, Stream},
         streambuilder::StreamBuilder,
+        struct_::{Core as _, Struct},
         symbol::{Core as _, Symbol},
         vector::{Core as _, Vector},
     },
