+++ src/mu/core/types.rs
@@ -21,8 +21,8 @@ use {
             float::{Core as _, Float},
             function::{Core as _, Function},
             map::{Core as _, Map},
-            r#struct::{Core as _, Struct},
             stream::{Core as _, Stream},
+            struct_::{Core as _, Struct},
             symbol::{Core as _, Symbol},
             vecimage::{TypedVec, VecType},
             vector::{Core as _, Vector},
