+++ src/mu/lib.rs
@@ -51,6 +51,7 @@ extern crate lazy_static;
 extern crate modular_bitfield;
 
 mod allocators;
+mod async_;
 mod core;
 mod system;
 mod types;
